apps/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-mptt==0.16.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
